fixes #27
The id that was being passed to the resource controller was the ID of
the SNOP and NOT the resource. I have changed the route back to accept a
domain_id and resource_id. This way, we are able to explicitly supply
the resource id and there is no confusion with the snop id.