Adding a few search story integration tests
There's a fairly complex problem here somewhere. One of the tests fails
due to what I believe is an issue with indexing. I've left a large
comment in the code for future test writing. I'm beginning to waste
WAYYY too much time on non necessary functionality and will be leaving
this alone for now. Later we can come back to it if there's time.