Cleaning up the resource page.
There was no reason to have a new route for the AJAX call. Since all we
were doing was just getting fresh information, we just needed to do an
AJAX GET on the Resource#show action. We can just check in there now if
the call is JS or HTML and render appropriately. This allows us to
remove some duplication of work (by having nearly identical show and
snop_show actions)