Nasty problem with routes fixed.
When you pass just a user to the user_path, it knows how to extract the
id, so you can pass a user instance INSTEAD of just the ID. However,
when you pass a snop along with the user, then you have to explicitly
pass in the user ID, or it wont know what to do with the user. Phew!