Removing everything that would run on document ready from application.js.
That stuff should go inside of the jQuery -> stuff in the CoffeeScript, so each page can do it differently.
Got domain table working properly.
Removing margin from the snop title link in the list view, and removing the color specification (it should be the default blue).
Moved the list flow buttons into a partial so we can render them from whatever pages we want.