Adding in the Yahoo reset locally, because the web one seemed to be taking a while to load.
Got the notice working properly, although it's not in the same spot as in the designs. The problem is that, if you were going into a list view, for example, the text would sit overtop. A nice thing to try would be to remove the flash message on every AJAX call, but I'm not sure how to do that...
Also fixing some alignment issues. This is a real pain trying to get the next/prev buttons going properly.