Switching the container padding by adding and removing a class, rather than by changing the CSS directly.
Hiding and showing the entire sidenav container, rather than all of the sidenav items themselves.
Setting up table properly for the user page.
Aligning a lot of the items better in the CSS.
Removing links in the user page list view content.
Making the subtitles on the list view pages more consistent (Snops from ...)
Adding a sidenav to the user show page.
Adding a container around all of the sidenav snops (sidenav_container) to assist in hiding/showing.