reloadClickHandlers has to take place in jQuery section to avoid intrusive JS in the views.
Now intelligently passing a snop list to the next/prev button partial so that it knows whether to enable or disable them based on which snop is current.
No longer yielding to tools_short directly from the application.html.erb -- this wasn't doing anything.
Moving the h2 subtitle to immediately before the all_views partial so that it appears in the correct spot on the page.