Fixing up search table so it matches the others.
Didn't need this browse_view parameter that was being set -- we just check the params in the partials.
Adding a new partial that basically renders any pages that involve lists of snops (including next/prev buttons, snop flow/list view, etc.)
Moving the sign up button directly into the application page -- we'll want it to appear everywhere if the user isn't signed in.