Fixes #116.
Switching name for left "list" sidenav to "tools". It's used on multiple pages, not just in snop list.
Made the drawing of the sidenav dependent upon a local rather than checking params[:browse_view]. This was necessary because we can't set the params inside the devise controllers. By just passing a local, we can force it to show if we want to.
Adding the tools sidenav to a bunch of pages. "Edit your settings", "Create a snop", and all user categories pages (create, index, edit/rename).