Two integration test errors.
One is expected, because we can't snop about a resource while looking at it, but that's a test.
The second is because after confirmation of an account, it seems devise tries to redirect to /users, but that path doesn't exist. We probably want to just set up a redirect or something at that route. Alternatively you can override the devise controller, but that's a bit uglier.